+++ src/components/feature/cookie-record/cookie-log-list/cookie-log-text/hint-icon/index.ts
