+++ src/components/feature/cookie-record/cookie-log-list/cookie-log-text/index.ts
