+++ src/components/feature/cookie-record/index.ts
@@ -1,0 +1,1 @@
+export * from './cookie-log-list'
